requirements.txt
alembic==1.17.1
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
bidict==0.23.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
granian==2.5.7
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==25.0
platformdirs==4.5.0
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
python-engineio==4.12.3
python-multipart==0.0.20
python-socketio==5.14.3
redis==7.0.1
reflex==0.8.18
reflex-hosting-cli==0.1.59
requests==2.32.5
rich==14.2.0
simple-websocket==1.1.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
watchfiles==1.1.1
wrapt==2.0.1
wsproto==1.2.0
pyspellchecker==0.8.3
